job-manager,sched,flux-simulator: add `.quiescent` services
The simulator can now send a `job-manager.quiescent` request, which will
only be responded to when the entire system has quiesced (i.e., in the
absence of new events/requests, the system will make no further changes
- such as allocating or freeing jobs).

The job-manager then sends its own `quiescent` request to the scheduler
along with every alloc request. It will only respond to the simulator's
request after its own request to the scheduler is responded to.  In the
future, this protocol will be expanded to include the exec and depend
modules.