python: `submit_async` returns a Future and `get_id` uses `wait_for`
`submit_async` used to return an opaque, c future pointer. Wrap it in a
python Future object instead so that users can leverage all Future
methods.

An error occurs if the submission future is not fulfilled when
`submit_get_id` is called.  Add a `wait_for` to ensure that the future
is fulfilled before getting the id.