scripts/check-format: update to optionally take files as arguments
If no arguments are provided, it will traverse the entire source tree
looking for python files to format.  If arguments are provided, only
those files are considered for checking.  The script will filter out any
non-python files from the arguments list before passing them to `black`.

Useful for git pre-commit scripts, which can produce a list of files
that changed in a given commit.  This results in a much lighterweight
operation than traversing the entire source tree and opening every file.